feat: 그룹 생성

+++ frontend/ARchive/app/src/main/java/com/droidblossom/archive/data/dto/group/request/CreateGroupRequestDto.kt
@@ -5,5 +5,5 @@ data class CreateGroupRequestDto(
     val groupDirectory: String,
     val groupImage: String,
     val groupName: String,
-    val targetIds: List<Int>
+    val targetIds: List<Long>
 )
